# docs/swagger.yaml
basePath: /
definitions:
  coding-challange_internal_model.APIErrorDetail:
    properties:
      code:
        type: integer
      details:
        items:
          $ref: '#/definitions/coding-challange_internal_model.ErrorDetail'
        type: array
      message:
        type: string
      type:
        type: string
    type: object
  coding-challange_internal_model.APIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/coding-challange_internal_model.APIErrorDetail'
      meta:
        $ref: '#/definitions/coding-challange_internal_model.Meta'
    type: object
  coding-challange_internal_model.APIResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/coding-challange_internal_model.Meta'
    type: object
  coding-challange_internal_model.ErrorDetail:
    properties:
      field:
        type: string
      issue:
        type: string
    type: object
  coding-challange_internal_model.Example:
    properties:
      explanation:
        type: string
      input:
        type: string
      output:
        type: string
    type: object
  coding-challange_internal_model.Hint:
    properties:
      content:
        type: string
      level:
        type: integer
      title:
        type: string
    type: object
  coding-challange_internal_model.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  coding-challange_internal_model.Parameter:
    properties:
      description:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  coding-challange_internal_model.Problem:
    properties:
      category:
        type: string
      constraints:
        items:
          type: string
        type: array
      description:
        type: string
      difficulty:
        type: string
      examples:
        items:
          $ref: '#/definitions/coding-challange_internal_model.Example'
        type: array
      function_name:
        type: string
      hints:
        items:
          $ref: '#/definitions/coding-challange_internal_model.Hint'
        type: array
      id:
        type: string
      parameters:
        items:
          $ref: '#/definitions/coding-challange_internal_model.Parameter'
        type: array
      return_type:
        type: string
      space_complexity_hint:
        type: string
      tags:
        items:
          type: string
        type: array
      template:
        type: string
      test_cases:
        items:
          $ref: '#/definitions/coding-challange_internal_model.TestCase'
        type: array
      time_complexity_hint:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/coding-challange_internal_model.ProblemType'
    type: object
  coding-challange_internal_model.ProblemSummary:
    properties:
      category:
        type: string
      difficulty:
        type: string
      id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  coding-challange_internal_model.ProblemType:
    enum:
    - function
    - main
    type: string
    x-enum-varnames:
    - ProblemTypeFunction
    - ProblemTypeMain
  coding-challange_internal_model.TestCase:
    properties:
      description:
        type: string
      expected: {}
      input:
        type: string
      params:
        items: {}
        type: array
    type: object
...
host: localhost:9100
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: API for the Coding Challenge platform — list problems, get details,
    validate code, and retrieve hints.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
  title: Coding Challenge API
  version: 1.1.0
paths:
  /api/problems:
    get:
      consumes:
      - application/json
      description: Get a paginated list of coding challenge problems with optional
        filters
      parameters:
      - description: Filter by difficulty (easy, medium, hard)
        in: query
        name: difficulty
        type: string
      - description: Filter by category (e.g. array, string)
        in: query
        name: category
        type: string
      - description: Filter by type (function, main)
        in: query
        name: type
        type: string
      - description: Filter by tags (comma-separated)
        in: query
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/coding-challange_internal_model.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/coding-challange_internal_model.ProblemSummary'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/coding-challange_internal_model.APIErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/coding-challange_internal_model.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/coding-challange_internal_model.APIErrorResponse'
      summary: List all problems
      tags:
      - problems
  /api/problems/{id}:
    get:
      consumes:
      - application/json
      description: Get full details of a specific coding challenge problem by its
        ID
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/coding-challange_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/coding-challange_internal_model.Problem'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/coding-challange_internal_model.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/coding-challange_internal_model.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/coding-challange_internal_model.APIErrorResponse'
      summary: Get a problem by ID
      tags:
      - problems
...
schemes:
- http
- https
swagger: "2.0"